% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpdb-package.R
\docType{package}
\name{bnpdb}
\alias{bnpdb-package}
\alias{bnpdb}
\title{The \code{bnpdb} Package}
\description{
\code{bnpdb} provides tools for Bayesian nonparametric modeling of time-to-event data using
Dirichlet process mixture models (DPMMs), with support for dynamic borrowing of information
from external data sources. These models allow flexible modeling of complex survival data
with latent subpopulations and nonparametric prior distributions on the space of densities.
}
\details{
Features of the package include:
\itemize{
\item Nonparametric regression using DPMMs.
\item Posterior summaries for conditional and marginal survival, hazard, and RMST functions.
\item Integration of internal and external data sources with a latent exchangeability framework.
\item Tools for visualizing and comparing model-based inferences.
}
}
\references{
Blackwell, D. and MacQueen, J. (1973). Ferguson distributions via Polya urn schemes. \emph{Annals of Statistics}, \strong{1}, 353–355.
Ferguson, T. S. (1974). Prior distributions on the spaces of probability measures. \emph{Annals of Statistics}, \strong{2}, 615–629.
Sethuraman, J. (1994). A constructive definition of Dirichlet priors. \emph{Statistica Sinica}, \strong{2}, 639–650.
Lo, A. Y. (1984). On a class of Bayesian nonparametric estimates I: Density estimates. \emph{Annals of Statistics}, \strong{12}, 351–357.
Escobar, M. D. (1994). Estimating normal means with a Dirichlet process prior. \emph{Journal of the American Statistical Association}, \strong{89}, 268–277.
Escobar, M. D. and West, M. (1995). Bayesian density estimation and inference using mixtures. \emph{Journal of the American Statistical Association}, \strong{90}, 577–588.
Neal, R. M. (2000). Markov chain sampling methods for Dirichlet process mixture models. \emph{Journal of Computational and Graphical Statistics}, \strong{9}, 249–265.
Ishwaran, H. and James, L. F. (2001). Gibbs sampling methods for stick-breaking priors. \emph{Journal of the American Statistical Association}, \strong{96}, 161–173.
Ishwaran, H. and James, L. F. (2002). Approximate Dirichlet process computing in finite normal mixtures: smoothing and prior information. \emph{Journal of Computational and Graphical Statistics}, \strong{11}, 508–532.
Alt, E. M., Chang, X., Jiang, X., Liu, Q., Mo, M., Xia, H. A., and Ibrahim, J. G. (2024). LEAP: the latent exchangeability prior for borrowing information from historical data. \emph{Biometrics}, \strong{80}(3), ujae083.
}
