% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_print_and_summarize.R
\name{print.bnpdb_tte_conditional_summary}
\alias{print.bnpdb_tte_conditional_summary}
\title{Print Method for 'bnpdb_tte_conditional_summary' Objects}
\usage{
\method{print}{bnpdb_tte_conditional_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bnpdb_tte_conditional_summary}, typically returned by
\code{\link{summary.bnpdb_tte_conditional}}.}

\item{...}{Additional arguments passed to the print method.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Prints summary statistics of posterior draws for parameters from a conditional
Bayesian nonparametric time-to-event model. Relies on the default method for
objects of class \code{posterior::draws_summary}.
}
\seealso{
\code{\link{summary.bnpdb_tte_conditional}}, \code{\link[posterior]{print.draws_summary}}
}
\keyword{internal}
