% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_plot.R
\name{tte_plot}
\alias{tte_plot}
\title{Plot time-to-event posterior}
\usage{
tte_plot(
  marginal_samples,
  point_estimate = "mean",
  ci_prob = 0.95,
  quantity = c("survival", "hazard", "rmst")
)
}
\arguments{
\item{marginal_samples}{output from calling \code{\link{tte_marginalize}}}

\item{point_estimate}{quantity to use as the point estimate when plotting}

\item{ci_prob}{credible interval bounds to plot; if \code{NULL}, only plots the point estimate}

\item{quantity}{a \code{character} giving which posterior quantities to plot}
}
\value{
a list of posterior function plots
}
\description{
Plot the posterior for time-to-event data
}
\examples{
##  Generate current and historical data sets
set.seed(741)
n     <- 100
n0    <- 100
N     <- n + n0
pexch <- 0.80
beta  <- cbind(c(1, 1, 0), c(-1, -1, 0))
sigma <- c(0.5, 0.75)
a     <- rbinom(N, 1, 0.5)      ##  tretment indicator
x     <- rbinom(N, 1, 0.5)      ##  binary covariate
eps0  <- rbinom(n0, 1, pexch)   ##  exchangeability indicator
eps   <- c(rep(1, n), 2 - eps0) ##  1 = exch; 2 = unexch
X     <- cbind(1, a, x)
Mu    <- unlist( 
  lapply(1:N, function(i) { as.numeric(X[i, ] \%*\% beta[, eps[i]] ) } ) 
)
logt     <- rnorm(N, Mu, sigma[eps])
logc     <- rnorm(N, mean(Mu) + 0.50, sd = 0.25)
event    <- logt <= logc
logy     <- ifelse(event, logt, logc)
dat      <- data.frame(y = exp(logy), event = event, a = a, x = x)
curdata  <- dat[1:n, ]
histdata <- dat[-(1:n), ]

##  Times for marginalization
tau   = max(curdata$y[curdata$event == 1])
times = seq(0.01, tau, length.out = 5)

##  Fit treatment only (with borrowing)
fit.strataonly = tte_reg_bnp(
  survival::Surv(y, event) ~ a, data = curdata, K = 5
  , external_data = histdata, K_unexch = 5
  , nburnin = 0, nsamples = 100
)

##  Marginalize strata only (no Bayesian bootstrap)
marg.strataonly = tte_marginalize(
  fit.strataonly, curdata, stratum_var = 'a'
  , hazard_times = times, log_hazard = TRUE
  , survival_times = times, log_survival = FALSE
)

#' plot
suppressWarnings({
  plotlist = tte_plot(marg.strataonly)
})

}
